photos-builder/requirements.txt
azure-functions
azure-storage-blob>=12.19.0
azure-kusto-data>=4.3.1
azure-kusto-ingest>=4.3.1
python-dateutil>=2.8.2
